+++ src/Controller/Game.cpp
@@ -10,22 +10,22 @@
 #include "Action.h"
 
 Game * Game::init(const char * filename) {
-	auto * smallTriangle1 = new Piece<double>(1, sf::Color(255, 100, 0),
-											  0.0, 400.0, 100.0, 300.0, 200.0, 400.0);
-	auto * smallTriangle2 = new Piece<double>(1, sf::Color(255, 255, 0),
-											  200.0, 200.0, 300.0, 100.0, 300.0, 300.0);
-	auto * mediumTriangle = new Piece<double>(1, sf::Color(0, 255, 0),
-											  200.0, 400.0, 400.0, 200.0, 400.0, 400.0);
-	auto * largeTriangle1 = new Piece<double>(1, sf::Color(150, 0, 100),
-											  0.0, 0.0, 200.0, 200.0, 0.0, 400.0);
-	auto * largeTriangle2 = new Piece<double>(1, sf::Color(255, 0, 150),
-											  0.0, 0.0, 400.0, 0.0, 200.0, 200.0);
-	auto * square = new Piece<double>(2, sf::Color(0, 0, 255),
-											  100.0, 300.0, 200.0, 200.0, 200.0, 400.0,
-											  200.0, 200.0, 300.0, 300.0, 200.0, 400.0);
-	auto * parallelogram = new Piece<double>(2, sf::Color(0, 255, 255),
-											  300.0, 100.0, 400.0, 0.0, 400.0, 200.0,
-											  300.0, 100.0, 400.0, 200.0, 300.0, 300.0);
+	auto * smallTriangle1 = new Piece<int>(1, sf::Color(255, 100, 0),
+											  0, 400, 100, 300, 200, 400);
+	auto * smallTriangle2 = new Piece<int>(1, sf::Color(255, 255, 0),
+											  200, 200, 300, 100, 300, 300);
+	auto * mediumTriangle = new Piece<int>(1, sf::Color(0, 255, 0),
+											  200, 400, 400, 200, 400, 400);
+	auto * largeTriangle1 = new Piece<int>(1, sf::Color(150, 0, 100),
+											  0, 0, 200, 200, 0, 400);
+	auto * largeTriangle2 = new Piece<int>(1, sf::Color(255, 0, 150),
+											  0, 0, 400, 0, 200, 200);
+	auto * square = new Piece<int>(2, sf::Color(0, 0, 255),
+											  100, 300, 200, 200, 200, 400,
+											  200, 200, 300, 300, 200, 400);
+	auto * parallelogram = new Piece<int>(2, sf::Color(0, 255, 255),
+											  300, 100, 400, 0, 400, 200,
+											  300, 100, 400, 200, 300, 300);
 
 	Game * game = new Game;
     game->pieces.push_back(largeTriangle1);
@@ -36,14 +36,14 @@ Game * Game::init(const char * filename) {
 	game->pieces.push_back(smallTriangle1);
 	game->pieces.push_back(smallTriangle2);
 
-	game -> goal = Piece<double>::createPiece(filename);
+	game -> goal = Piece<int>::createPiece(filename);
 
 	// boutons :
 	// ???
 	return game;
 }
 
-Game::Game() : selected(nullptr), relativePos(Point(0., 0.)) {
+Game::Game() : selected(nullptr), relativePos(Point(0, 0)) {
 }
 
 void Game::draw(sf::RenderWindow& window) {
@@ -57,7 +57,7 @@ void Game::draw(sf::RenderWindow& window) {
 
 /* Select a piece when a left click is performed by the user
 */
-void Game::select(const Point<double> & event) {
+void Game::select(const Point<int> & event) {
 	selected = nullptr;
 	for (auto s : pieces) {
 		if (s->isClicked(event)){
@@ -67,7 +67,7 @@ void Game::select(const Point<double> & event) {
 	}
 }
 
-void Game::deselect(const Point<double> & event) {
+void Game::deselect(const Point<int> & event) {
 	if (selected != nullptr) {
 		//selected->centralize(event, relativePos);
 		magnetize();
@@ -91,22 +91,22 @@ Game::~Game() {
 }
 
 void Game::save() {
-	std::vector<Point<double>> points;
-	for_each(pieces.cbegin(), pieces.cend(), [&points](Shape<double> * s) {
-		vector<Point<double>> shapePoints = s -> getPoints();
+	std::vector<Point<int>> points;
+	for_each(pieces.cbegin(), pieces.cend(), [&points](Shape<int> * s) {
+		vector<Point<int>> shapePoints = s -> getPoints();
 		points.insert(points.cend(), shapePoints.cbegin(), shapePoints.cend());
 	});
 
-	Point<double>::normalize(points);
+	Point<int>::normalize(points);
 	FileUtils::writeFile(points, "levels/save.txt");
 }
 
 void Game::magnetize() {
     double dist;
     double minDist = DBL_MAX;
-    std::vector<Point<double>> points(2);
-    std::vector<Point<double>> minPoints(2);
-    Point<double> translation;
+    std::vector<Point<int>> points(2);
+    std::vector<Point<int>> minPoints(2);
+    Point<int> translation;
     // Get the shortest distance between two points from each shape
     for(auto& piece : pieces) {
         if(piece != selected) {
